pyatls: restructure validators into hierarchy
