+++ python-package/src/atls/validators/azure/aas/shared.py
@@ -1,6 +1,6 @@
 from typing import List
 
-AZ_AAS_GLOBAL_JKUS: List[str] = [
+PUBLIC_JKUS: List[str] = [
     "https://sharedcus.cus.attest.azure.net/certs",
     "https://sharedeus.eus.attest.azure.net/certs",
     "https://sharedeus2.eus2.attest.azure.net/certs",
